[strategies] Relocate CS-specific winding strategies.
